Fold embedding-space resolution into EmbedderRegistry and embed_samples
Remove default_embedding_space. EmbedderRegistry now owns embedders and the
built-in bootstrap defaults with no database access: register_embedder and
bootstrap_space return an EmbeddingSpaceSpec. embed_samples owns the database
steps via ensure_default_model: read the collection's default model, else ask
the registry for a bootstrap space and persist it, then request the embedder
from the registry.
